requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
PyGithub==2.1.1
playwright==1.40.0
openrouter-sdk==0.1.0
moviepy==1.0.3
pillow==10.1.0
werkzeug==3.0.1